# apps/api/src/swagger/swagger_ui.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "POP Api",
		"description": "Documentation sur les webservices de l'application Api"
	},
	"host": "localhost:3000",
	"basePath": "/",
...
	"schemes": ["http"],
	"securityDefinitions": {
		"apiKeyAuth": {
			"type": "apiKey",
			"in": "header",
			"name": "X-API-KEY",
			"description": "any description..."
		}
	},
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/memoire/{ref}": {
			"put": {
				"tags": ["Mémoire"],
				"description": "Modification de la notice Mémoire",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								},
								"updateMode": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			},
			"get": {
				"tags": ["Mémoire"],
				"description": "Retourne les informations de la notice Mémoire",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "Référence de la notice Mémoire"
					}
				],
				"responses": {
					"200": {
						"description": "Récupération des informations avec succés",
						"schema": {
							"$ref": "#/definitions/GetMemoire"
						}
					},
					"404": {
						"description": "Document non trouvé",
						"schema": {
							"type": "object",
							"properties": {
								"success": {
									"type": "boolean",
									"example": false
								},
								"msg": {
									"type": "string",
									"example": "Document introuvable"
								}
							},
							"xml": {
								"name": "main"
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Mémoire"],
				"description": "Suppression de la notice Mémoire",
				"parameters": [
					{
						"name": "ref",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
		"/memoire": {
			"post": {
				"tags": ["Mémoire"],
				"description": "Création de la notice Mémoire",
				"parameters": [
					{
						"name": "body",
						"in": "body",
						"schema": {
							"type": "object",
							"properties": {
								"notice": {
									"example": "any"
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"500": {
						"description": "Internal Server Error"
					}
				}
			}
		},
...
	},
	"definitions": {
...
		"GetMemoire": {
			"type": "object",
			"properties": {
				"PRODUCTEUR": {
					"type": "string",
					"example": ""
				},
				"BASE": {
					"type": "string",
					"example": ""
				},
				"CONTIENT_IMAGE": {
					"type": "string",
					"example": "oui"
				},
				"POP_FLAGS": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"POP_IMPORT": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"TOUT": {
					"type": "string",
					"example": ""
				},
				"ADRESSE": {
					"type": "string",
					"example": ""
				},
				"AUTOEU": {
					"type": "string",
					"example": ""
				},
				"AUTG": {
					"type": "string",
					"example": ""
				},
				"AUTP": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"AUTOR": {
					"type": "string",
					"example": ""
				},
				"AUTTI": {
					"type": "string",
					"example": ""
				},
				"COM": {
					"type": "string",
					"example": ""
				},
				"DOM": {
					"type": "string",
					"example": ""
				},
				"EDIF": {
					"type": "string",
					"example": ""
				},
				"EXPO": {
					"type": "string",
					"example": ""
				},
				"JDATPV": {
					"type": "string",
					"example": ""
				},
				"LIEUCOR": {
					"type": "string",
					"example": ""
				},
				"COTECOR": {
					"type": "string",
					"example": ""
				},
				"LIEUCTI": {
					"type": "string",
					"example": ""
				},
				"COTECTI": {
					"type": "string",
					"example": ""
				},
				"LIEUCP": {
					"type": "string",
					"example": ""
				},
				"COTECP": {
					"type": "string",
					"example": ""
				},
				"LEG": {
					"type": "string",
					"example": ""
				},
				"OBJT": {
					"type": "string",
					"example": ""
				},
				"OBS": {
					"type": "string",
					"example": ""
				},
				"OBSOR": {
					"type": "string",
					"example": ""
				},
				"OBSTI": {
					"type": "string",
					"example": ""
				},
				"PAYS": {
					"type": "string",
					"example": ""
				},
				"PUBLI": {
					"type": "string",
					"example": ""
				},
				"TIREDE": {
					"type": "string",
					"example": ""
				},
				"ROLE": {
					"type": "string",
					"example": ""
				},
				"PRECOR": {
					"type": "string",
					"example": ""
				},
				"SERIE": {
					"type": "string",
					"example": ""
				},
				"THEATRE": {
					"type": "string",
					"example": ""
				},
				"TITRE": {
					"type": "string",
					"example": ""
				},
				"DMAJ": {
					"type": "string",
					"example": ""
				},
				"DMIS": {
					"type": "string",
					"example": ""
				},
				"IDPROD": {
					"type": "string",
					"example": ""
				},
				"NUMCD": {
					"type": "string",
					"example": ""
				},
				"NUMF": {
					"type": "string",
					"example": ""
				},
				"INSEE": {
					"type": "string",
					"example": ""
				},
				"NVD": {
					"type": "string",
					"example": ""
				},
				"MARQ": {
					"type": "string",
					"example": ""
				},
				"PLOC": {
					"type": "string",
					"example": ""
				},
				"ACQU": {
					"type": "string",
					"example": ""
				},
				"ADPHOT": {
					"type": "string",
					"example": ""
				},
				"AIRE": {
					"type": "string",
					"example": ""
				},
				"ANUMP": {
					"type": "string",
					"example": ""
				},
				"COPY": {
					"type": "string",
					"example": ""
				},
				"COULEUR": {
					"type": "string",
					"example": ""
				},
				"DPT_LETTRE": {
					"type": "string",
					"example": ""
				},
				"DATIMM": {
					"type": "string",
					"example": ""
				},
				"DATOEU": {
					"type": "string",
					"example": ""
				},
				"DATPV": {
					"type": "string",
					"example": ""
				},
				"DATOR": {
					"type": "string",
					"example": ""
				},
				"DATTI": {
					"type": "string",
					"example": ""
				},
				"DATG": {
					"type": "string",
					"example": ""
				},
				"DATD": {
					"type": "string",
					"example": ""
				},
				"DIFF": {
					"type": "string",
					"example": ""
				},
				"DPT": {
					"type": "string",
					"example": ""
				},
				"EDIARCH": {
					"type": "string",
					"example": ""
				},
				"ECH": {
					"type": "string",
					"example": ""
				},
				"FORMAT": {
					"type": "string",
					"example": ""
				},
				"FORMATOR": {
					"type": "string",
					"example": ""
				},
				"FORMATTI": {
					"type": "string",
					"example": ""
				},
				"LBASE": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"WEB": {
					"type": "string",
					"example": ""
				},
				"LIB": {
					"type": "string",
					"example": ""
				},
				"LOCA": {
					"type": "string",
					"example": ""
				},
				"LIEUORIG": {
					"type": "string",
					"example": ""
				},
				"MCGEO": {
					"type": "string",
					"example": ""
				},
				"MCL": {
					"type": "string",
					"example": ""
				},
				"MENTIONS": {
					"type": "string",
					"example": ""
				},
				"MENTOR": {
					"type": "string",
					"example": ""
				},
				"MENTTI": {
					"type": "string",
					"example": ""
				},
				"MCPER": {
					"type": "string",
					"example": ""
				},
				"VUECD": {
					"type": "string",
					"example": ""
				},
				"NUMAUTP": {
					"type": "string",
					"example": ""
				},
				"NUMVERS": {
					"type": "string",
					"example": ""
				},
				"ANUMOR": {
					"type": "string",
					"example": ""
				},
				"NUMOR": {
					"type": "string",
					"example": ""
				},
				"NUMP": {
					"type": "string",
					"example": ""
				},
				"ANUMTI": {
					"type": "string",
					"example": ""
				},
				"NUMTI": {
					"type": "string",
					"example": ""
				},
				"RENV": {
					"type": "string",
					"example": ""
				},
				"REG": {
					"type": "string",
					"example": ""
				},
				"SENS": {
					"type": "string",
					"example": ""
				},
				"SCLE": {
					"type": "string",
					"example": ""
				},
				"SUP": {
					"type": "string",
					"example": ""
				},
				"TECHN": {
					"type": "string",
					"example": ""
				},
				"TECHOR": {
					"type": "string",
					"example": ""
				},
				"TECHTI": {
					"type": "string",
					"example": ""
				},
				"CINEPROD": {
					"type": "string",
					"example": ""
				},
				"TYP": {
					"type": "string",
					"example": "AREF"
				},
				"TYPDOC": {
					"type": "string",
					"example": ""
				},
				"TYPEIMG": {
					"type": "string",
					"example": ""
				},
				"TYPSUPP": {
					"type": "string",
					"example": ""
				},
				"VIDEO": {
					"type": "string",
					"example": ""
				},
				"LBASE2": {
					"type": "string",
					"example": ""
				},
				"LEG2": {
					"type": "string",
					"example": ""
				},
				"REFIM": {
					"type": "string",
					"example": ""
				},
				"REFIMG": {
					"type": "string",
					"example": ""
				},
				"REFJOC": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"REFMUS": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"MOSA": {
					"type": "string",
					"example": ""
				},
				"SITE": {
					"type": "string",
					"example": ""
				},
				"NUMSITE": {
					"type": "string",
					"example": ""
				},
				"NUMOP": {
					"type": "string",
					"example": ""
				},
				"CHRONO": {
					"type": "string",
					"example": ""
				},
				"REPRO": {
					"type": "string",
					"example": ""
				},
				"STRUCT": {
					"type": "string",
					"example": ""
				},
				"SUJET": {
					"type": "string",
					"example": ""
				},
				"TICO": {
					"type": "string",
					"example": ""
				},
				"NUMI": {
					"type": "string",
					"example": ""
				},
				"LIEU": {
					"type": "string",
					"example": ""
				},
				"WADRS": {
					"type": "string",
					"example": ""
				},
				"CONTACT": {
					"type": "string",
					"example": ""
				},
				"EMET": {
					"type": "string",
					"example": ""
				},
				"NUM": {
					"type": "string",
					"example": ""
				},
				"IMG": {
					"type": "string",
					"example": ""
				},
				"WCOM": {
					"type": "string",
					"example": ""
				},
				"LIENS": {
					"type": "string",
					"example": ""
				},
				"LAUTP": {
					"type": "string",
					"example": ""
				},
				"TRL": {
					"type": "string",
					"example": ""
				},
				"DENO": {
					"type": "string",
					"example": ""
				},
				"AUT": {
					"type": "string",
					"example": ""
				},
				"AUTR": {
					"type": "string",
					"example": ""
				},
				"DOC": {
					"type": "string",
					"example": ""
				},
				"NEGPOS": {
					"type": "string",
					"example": ""
				},
				"NOMSN": {
					"type": "string",
					"example": ""
				},
				"_id": {
					"type": "string",
					"example": ""
				},
				"REF": {
					"type": "string",
					"example": ""
				},
				"__v": {
					"type": "number",
					"example": 1
				},
				"MEMOIRE": {
					"type": "array",
					"example": [],
					"items": {}
				},
				"POP_COMMENTAIRES": {
					"type": "array",
					"example": [""],
					"items": {
						"type": "string"
					}
				},
				"HISTORIQUE": {
					"type": "array",
					"items": {
						"type": "object",
						"properties": {
							"_id": {
								"type": "string",
								"example": ""
							},
							"nom": {
								"type": "string",
								"example": ""
							},
							"prenom": {
								"type": "string",
								"example": ""
							},
							"email": {
								"type": "string",
								"example": ""
							},
							"date": {
								"type": "string",
								"example": ""
							},
							"updateMode": {
								"type": "string",
								"example": ""
							}
						}
					}
				}
			}
		},
...
	}
}